EF: Build errors found by VC12.

+++ iModelCore/ecobjects/test/NonPublishedScenario/SchemaTests.cpp
@@ -32,7 +32,7 @@ TEST_F(SchemaTest, ShouldBeAbleToIterateOverECClassContainer)
         {
         ECClassCP ecClass = *cit;
         WString name = ecClass->GetName();
-        wprintf(L"ECClass=0x%x, name=%s\n", ecClass, name.c_str());
+        wprintf(L"ECClass=0x%x, name=%s\n", (uintptr_t)ecClass, name.c_str());
         count++;
         }
     ASSERT_EQ(2, count);
@@ -40,7 +40,7 @@ TEST_F(SchemaTest, ShouldBeAbleToIterateOverECClassContainer)
     FOR_EACH (ECClassCP ecClass, container)
         {
         WString name = ecClass->GetName();
-        wprintf(L"ECClass=0x%x, name=%s\n", ecClass, name.c_str());
+        wprintf(L"ECClass=0x%x, name=%s\n", (uintptr_t)ecClass, name.c_str());
         count++;
         }
     ASSERT_EQ(4, count);
@@ -67,7 +67,7 @@ TEST_F(SchemaTest, TestGetClassCount)
         {
         ECClassCP ecClass = *cit;
         WString name = ecClass->GetName();
-        wprintf(L"ECClass=0x%x, name=%s\n", ecClass, name.c_str());
+        wprintf(L"ECClass=0x%x, name=%s\n", (uintptr_t)ecClass, name.c_str());
         count++;
         }
     ASSERT_EQ(2, count);
@@ -75,7 +75,7 @@ TEST_F(SchemaTest, TestGetClassCount)
     FOR_EACH (ECClassCP ecClass, container)
         {
         WString name = ecClass->GetName();
-        wprintf(L"ECClass=0x%x, name=%s\n", ecClass, name.c_str());
+        wprintf(L"ECClass=0x%x, name=%s\n", (uintptr_t)ecClass, name.c_str());
         count++;
         }
     ASSERT_EQ(4, count);
